refactor(fe-api-routes): change api route directory structure
