+++ frontend/src/app.feature/dashboard.page-transition/component/BarPageTransition.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Bar, BarChart, Tooltip, XAxis, YAxis } from "recharts";
 import useQueryFn from "@/app.module/react-query/useQueryFn";
 import { BestPageTransition } from "@/app.feature/dashboard.entry/types/pageTransition";
-import { API_BEST_BARS_TRANSITION } from "@/app.module/constant/api/app.dashboard";
+import { API_BEST_BARS_TRANSITION } from "@/app.module/constant/api/app.dashboard/transition";
 import DashboardBackground from "@/app.components/dashboard/DashboardBackground";
 
 const BarPageTransition = () => {
