--- frontend/src/app.module/constant/api/app.dashboard.ts
@@ -1,3 +1,0 @@
-export const API_SANKEY_TRANSITION = '/event/page-transitions';
-export const API_BEST_BARS_TRANSITION = '/event/page-transition-best';
-export const API_DAILY_VIEW = '/overview/visitors';
