+++ frontend/src/app.module/constant/api/app.dashboard/overview.ts
@@ -1,0 +1,11 @@
+/* Overview Section */
+
+/**
+ * visitors view count data.
+ */
+export const API_DAILY_VIEW = "/overview/visitors";
+
+/**
+ * top stayed pages data.
+ */
+export const API_TOP_STAYED = "/overview/top-stayed";
