+++ frontend/src/app.module/constant/api/app.dashboard/transition.ts
@@ -1,0 +1,5 @@
+/* Page Transition */
+
+/* will be changed */
+export const API_SANKEY_TRANSITION = "/event/page-transitions";
+export const API_BEST_BARS_TRANSITION = "/event/page-transition-best";
